Back End/requirements.txt
Flask~=1.1.2
Werkzeug~=1.0.1
mysql-connector-python~=8.0.22
requests~=2.25.1